feat: add actions for register/login/logout
