+++ src/components/App/App.js
@@ -1,6 +1,11 @@
-import React from 'react';
+/* eslint-disable no-shadow */
+import React, { useEffect } from 'react';
 import { BrowserRouter as Router, Switch } from 'react-router-dom';
 import { Container } from 'reactstrap';
+import { connect } from 'react-redux';
+import { bindActionCreators } from 'redux';
+import PropTypes from 'prop-types';
+import { changeStatus } from '../../store/actions';
 import Header from '../Navigation';
 import Footer from '../../UI/Footer';
 import AuthContextProvider from '../AuthContextProvider';
@@ -8,7 +13,10 @@ import Routes from '../Routes';
 import AlertsContainer from '../AlertsContainer';
 import './app.scss';
 
-const App = () => {
+const App = ({ changeStatus }) => {
+  useEffect(() => {
+    changeStatus();
+  }, []);
   return (
     <AuthContextProvider>
       <Router>
@@ -25,4 +33,10 @@ const App = () => {
   );
 };
 
-export default App;
+App.propTypes = {
+  changeStatus: PropTypes.func.isRequired,
+};
+
+const mapDispatchToProps = (dispatch) => bindActionCreators({ changeStatus }, dispatch);
+
+export default connect(null, mapDispatchToProps)(App);
