+++ src/store/actions/actionTypes.js
@@ -16,6 +16,10 @@ const GET_USER_PROGRESS = 'GET_USER_PROGRESS';
 const DELETE_USER_PROGRESS = 'DELETE_USER_PROGRESS';
 const EDIT_USER_PROGRESS = 'EDIT_USER_PROGRESS';
 const ADD_USER_PROGRESS = 'ADD_USER_PROGRESS';
+const AUTH_LOG_IN = 'AUTH_LOG_IN';
+const CHANGE_AUTH_STATUS = 'CHANGE_AUTH_STATUS';
+const AUTH_LOG_OUT = 'AUTH_LOG_OUT';
+const AUTH_REGISTER = 'AUTH_REGISTER';
 
 export {
   FETCH_MEMBERS,
@@ -36,4 +40,8 @@ export {
   DELETE_USER_PROGRESS,
   EDIT_USER_PROGRESS,
   ADD_USER_PROGRESS,
+  AUTH_LOG_IN,
+  CHANGE_AUTH_STATUS,
+  AUTH_LOG_OUT,
+  AUTH_REGISTER,
 };
